add cast to remove compiler warning

+++ src/chunk_list.cpp
@@ -444,7 +444,7 @@ static void chunk_log(chunk_t *pc, const char *text)
       {
          chunk_log_msg(prev, log, " @ after");
       }
-      LOG_FMT(log, " stage is %d", cpd.unc_stage);
+      LOG_FMT(log, " stage is %u", static_cast<unsigned int>(cpd.unc_stage));
       log_func_stack_inline(log);
    }
 }
